+++ mainwindow.h
@@ -36,7 +36,6 @@
 #include "src/gui/dialogues/newtaildialog.h"
 #include "src/gui/dialogues/newpilotdialog.h"
 #include "src/classes/arunguard.h"
-#include "src/experimental/adatabase.h"
 #include "src/testing/atimer.h"
 
 QT_BEGIN_NAMESPACE
