+++ src/classes/aaircraftentry.h
@@ -1,10 +1,8 @@
 #ifndef AAIRCRAFTENTRY_H
 #define AAIRCRAFTENTRY_H
 
-#include "src/experimental/aentry.h"
-#include "src/experimental/decl.h"
-
-namespace experimental {
+#include "src/classes/aentry.h"
+#include "src/database/declarations.h"
 
 struct AAircraftEntry : public AEntry {
 public:
@@ -12,9 +10,7 @@ public:
     AAircraftEntry(const AAircraftEntry& te) = default;
     AAircraftEntry& operator=(const AAircraftEntry& te) = default;
     AAircraftEntry(int row_id);
-    AAircraftEntry(TableData table_data);
+    AAircraftEntry(RowData table_data);
 };
-
-} // namespace experimental
 
 #endif // AAIRCRAFTENTRY_H
