+++ src/gui/widgets/aircraftwidget.cpp
@@ -17,12 +17,12 @@
  */
 #include "aircraftwidget.h"
 #include "ui_aircraftwidget.h"
-
-// Debug Makro
-#define DEB(expr) \
-    qDebug() << __PRETTY_FUNCTION__ << "\t" << expr
-
-using namespace experimental;
+#include "src/gui/dialogues/newtaildialog.h"
+#include "src/classes/asettings.h"
+#include "src/database/adatabase.h"
+#include "src/classes/atailentry.h"
+#include "src/classes/aflightentry.h"
+#include "src/testing/adebug.h"
 
 AircraftWidget::AircraftWidget(QWidget *parent) :
     QWidget(parent),
