+++ src/gui/widgets/aircraftwidget.h
@@ -24,12 +24,6 @@
 #include <QDebug>
 #include <QTableView>
 
-#include "src/classes/asettings.h"
-#include "src/gui/dialogues/newtaildialog.h"
-#include "src/experimental/adatabase.h"
-#include "src/experimental/atailentry.h"
-#include "src/experimental/aflightentry.h"
-
 
 namespace Ui {
 class AircraftWidget;
