+++ src/gui/widgets/debugwidget.h
@@ -12,9 +12,9 @@
 #include "src/classes/adownload.h"
 #include "src/functions/areadcsv.h"
 
-#include "src/experimental/adatabase.h"
-#include "src/experimental/aentry.h"
-#include "src/experimental/apilotentry.h"
+#include "src/database/adatabase.h"
+#include "src/classes/aentry.h"
+#include "src/classes/apilotentry.h"
 
 #include "src/testing/abenchmark.h"
 #include "src/testing/atimer.h"
