+++ src/gui/widgets/logbookwidget.h
@@ -26,10 +26,6 @@
 #include <QMenu>
 #include <QTableView>
 
-#include "src/classes/asettings.h"
-#include "src/gui/dialogues/newflightdialog.h"
-#include "src/experimental/aflightentry.h"
-
 namespace Ui {
 class LogbookWidget;
 }
