+++ src/gui/widgets/settingswidget.cpp
@@ -19,6 +19,10 @@
 #include "ui_settingswidget.h"
 #include "src/testing/adebug.h"
 
+#include "src/classes/asettings.h"
+#include "src/database/adatabase.h"
+#include "src/classes/apilotentry.h"
+
 static const auto FIRSTNAME_VALID = QPair<QString, QRegularExpression> {
     "firstnameLineEdit", QRegularExpression("[a-zA-Z]+")};
 static const auto LASTNAME_VALID = QPair<QString, QRegularExpression> {
@@ -40,8 +44,6 @@ static const auto LINE_EDIT_VALIDATORS = QVector({FIRSTNAME_VALID, LASTNAME_VALI
                                            COMPANY_VALID,     EMPLOYEENR_VALID,
                                            PREFIX_VALID});
 
-using namespace experimental;
-
 SettingsWidget::SettingsWidget(QWidget *parent) :
     QWidget(parent),
     ui(new Ui::SettingsWidget)
@@ -58,7 +60,7 @@ SettingsWidget::SettingsWidget(QWidget *parent) :
     setupValidators();
 
     QObject::connect(themeGroup, QOverload<int>::of(&QButtonGroup::buttonClicked),
-                     this, &SettingsWidget::on_themeGroup_buttonClicked);
+                     this, &SettingsWidget::onThemeGroup_buttonClicked);
 }
 
 SettingsWidget::~SettingsWidget()
@@ -262,7 +264,7 @@ void SettingsWidget::on_prefixLineEdit_textChanged(const QString &arg1)
 /*
  * Misc Tab
  */
-void SettingsWidget::on_themeGroup_buttonClicked(int theme_id)
+void SettingsWidget::onThemeGroup_buttonClicked(int theme_id)
 {
     ASettings::write("main/theme", theme_id);
