+++ src/gui/widgets/settingswidget.h
@@ -24,9 +24,6 @@
 #include <QMessageBox>
 #include <QProcess>
 #include <QDebug>
-#include "src/classes/asettings.h"
-#include "src/experimental/adatabase.h"
-#include "src/experimental/apilotentry.h"
 
 namespace Ui {
 class SettingsWidget;
@@ -42,7 +39,7 @@ public:
 
 private slots:
 
-    void on_themeGroup_buttonClicked(int theme_id);
+    void onThemeGroup_buttonClicked(int theme_id);
     void on_aboutPushButton_clicked();
     void on_acSortComboBox_currentIndexChanged(int index);
     void on_acAllowIncompleteComboBox_currentIndexChanged(int index);
